* develop: (88 commits)
  Fix warnings
  Compilation warning fixes
  Fix for Intel 17
  ECKIT-341: cleanup
  ECKIT-341: fix
  ECKIT-341: test different linear algebra backends on spmm
  ECKIT-340: use __func__ from c++11
  Sandbox
  Python client and server json event
  Sandbox send JSON event
  Bring sandbox to eckit
  ECKIT-339 - Monitor reset
  Fix compile errors.
  Fix compile errors.
  Fix compilation for Cray 8.6.2
  Remove warning
  ECKIT-337: fix warning
  ECKIT-336: move more tests
  Compilation fix
  ECKIT-336: move more tests
  ...

+++ src/eckit/runtime/Main.cc
@@ -53,7 +53,7 @@ Main::Main(int argc, char** argv, const char* homeenv) :
 
     name_ = displayName_ = PathName(argv[0]).baseName(false);
 
-    for (size_t i = 1; i < size_t(argc); ++i) {
+    for (int i = 1; i < argc; ++i) {
 
         // Old style
         if (::strcmp(argv[i], "-debug") == 0) {
